Add and Modify
Add article
Modify ArticleToHtml.py

+++ ArticleToHtml/ArticleToHtml.py
@@ -1,5 +1,5 @@
 # -*- coding: GBK -*-
-articlePath_ = "..\\article\\origin\\UbuntuKylin下安装配置LAMP.txt"
+articleOriginPath_ = "..\\article\\origin\\"
 articleListPath_="..\\article\\articleList.xml"
 htmlDir_ = "..\\article\\html\\"
 
@@ -9,12 +9,48 @@ sys.path.append("UtilSp\Python")
 import os
 from fileSp import *
 from xmlSp import *
+from listSp import *
+
+def calculateConverArticleList(articleIdNameDict,articleOriginList):    
+    if  articleIdNameDict!=None:
+        waitConvertArticleList=[] 
+        for articleOrigin in articleOriginList:
+            articleOriginName=fetchFileName(articleOrigin,False)
+            hasExistInList=False
+            id=10001 
+            while str(id) in articleIdNameDict.keys():
+                articleName=articleIdNameDict[str(id)] 
+                if articleName==articleOriginName:                    
+                    hasExistInList=True                    
+                    break;               
+                id+=1
+            if hasExistInList:
+                continue
+            else:
+                if not (articleOrigin in waitConvertArticleList):
+                    waitConvertArticleList.append(articleOriginName)   
+        lastArticleId=10001+len(articleIdNameDict)-1         
+        return [lastArticleId,waitConvertArticleList]
+    else:
+        return None
+
+def fetchArticleOriginList(articleOriginPath):
+    filenameList=os.listdir(articleOriginPath)
+    return filenameList
 
 def fetchArticleIdList(artilceListPath):
     articleListXml=xmlSp()   
     xmlTree=articleListXml.fetchXmlNodeTree(artilceListPath)
     idList=articleListXml.fetchNodeValueList(xmlTree,'id')
     return idList
+
+def fetchArticleIdNameDict(artilceListPath):
+    articleListXml=xmlSp()   
+    xmlTree=articleListXml.fetchXmlNodeTree(artilceListPath)
+    idList=articleListXml.fetchNodeValueList(xmlTree,'id')
+    nameList=articleListXml.fetchNodeValueList(xmlTree,'name')
+    idNameDict=twoListToDict(idList,nameList)
+    return idNameDict
 
 def fetchLastArticleId(artilceListPath):
     articleIdList=fetchArticleIdList(artilceListPath)
@@ -25,16 +61,19 @@ def fetchLastArticleId(artilceListPath):
        lastId=int(articleIdList[idCount-1])
        return lastId
 
-def outHtml():
-    articleName=fetchFileName(articlePath_,False)
-    lastId=fetchLastArticleId(articleListPath_)
-    newId=lastId+1;
+def outHtml(newId,convertArticleName):    
+    #lastId=fetchLastArticleId(articleListPath_)
+    #newId=lastId+1;   
     htmlName=str(newId)+'.txt'
     htmlPath=htmlDir_+htmlName
-    articleFile = open(articlePath_,"r")
+    articleOriginPath=articleOriginPath_+convertArticleName
+    if not os.path.isfile(articleOriginPath):
+        print(convertArticleName+' is not file')
+        return
+    articleFile = open(articleOriginPath,"r")
     htmlFile = open(htmlPath,"w")
     htmlFile.write("<h3>")
-    htmlFile.write(articleName)
+    htmlFile.write(convertArticleName)
     htmlFile.write("</h3>\n")
     for line in articleFile:    
         htmlFile.write("<div>\n")
@@ -48,24 +87,46 @@ def outHtml():
     outParam={}
     outParam['htmlPath']=htmlPath
     outParam['articldId']=newId
-    outParam['articldName']=articleName
+    outParam['articldName']=convertArticleName
     return outParam
 
-def updateArticleList(articleId,articleName):
+def outHtmlBatch():   
+    articleIdNameDict=fetchArticleIdNameDict(articleListPath_)
+    articleOriginList=fetchArticleOriginList(articleOriginPath_) 
+    waitConvert=calculateConverArticleList(articleIdNameDict,articleOriginList)
+    if waitConvert!=None:
+        lastId=waitConvert[0]
+        convertArticleList=waitConvert[1]
+        #print(convertArticleList)
+        if  len(convertArticleList)>0:
+            newId=lastId+1
+            newArticleIdList=[]
+            for articleName in convertArticleList:                
+                outHtml(newId,articleName+".txt")
+                newArticleIdList.append(newId)                
+                newId+=1
+            updateArticleList(newArticleIdList,convertArticleList)                
+
+
+def updateArticleList(articleIdList,articleNameList):
     articleListXml = xmlSp()  
     nodeTree = articleListXml.fetchXmlNodeTree(articleListPath_)
     rootNode=articleListXml.fetchSingleNode(nodeTree,'articleList')
-    idNode=articleListXml.createChildNode('id',str(articleId))
-    nameNode=articleListXml.createChildNode('name',str(articleName))
-    parentNode=articleListXml.createChildNode('article','')
-    articleListXml.addNode(parentNode,idNode)
-    articleListXml.addNode(parentNode,nameNode)
-    articleListXml.addNode(rootNode,parentNode)
-    articleListXml.writeXml(nodeTree,articleListPath_)
+    for index in range(len(articleIdList)):
+        articleId=articleIdList[index]
+        articleName=articleNameList[index]
+        idNode=articleListXml.createChildNode('id',str(articleId))
+        nameNode=articleListXml.createChildNode('name',str(articleName))
+        parentNode=articleListXml.createChildNode('article','')
+        articleListXml.addNode(parentNode,idNode)
+        articleListXml.addNode(parentNode,nameNode)
+        articleListXml.addNode(rootNode,parentNode)
+        articleListXml.writeXml(nodeTree,articleListPath_)
 
 if __name__ == '__main__': 
-    outParam=outHtml()
-    print(articlePath_+" convert to "+outParam['htmlPath']+" OK!")
-    updateArticleList(outParam['articldId'],outParam['articldName'])
+    #convertArticleName=input("Please input article name that want to convert...\n")
+    outHtmlBatch()  
+    #print(articleOriginPath_+" convert to "+outParam['htmlPath']+" OK!")
+    #updateArticleList(outParam['articldId'],outParam['articldName'])
     os.system("PAUSE")
